refactor: 공용되는 스타일 컴포넌트 분리
