+++ app/src/components/form/CardCVCInput.tsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import styled from "@emotion/styled";
 import { ErrorMessage } from "./ErrorMessage";
+import { CardInput } from "../../style/CardInput";
 import { Validator } from "../../validators/CardValidator";
 
 export function CardCVCInputWrapper() {
@@ -35,7 +36,7 @@ export function CardCVCInputWrapper() {
   return (
     <CardCVCContainer>
       <CardCVCLabel htmlFor="card-cvc-input">CVC</CardCVCLabel>
-      <CardCVCInput
+      <CardInput
         type="text"
         maxLength={3}
         inputMode="numeric"
@@ -50,10 +51,6 @@ export function CardCVCInputWrapper() {
   );
 }
 
-type ErrorFlag = {
-  isError?: boolean;
-};
-
 const CardCVCContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -62,22 +59,4 @@ const CardCVCContainer = styled.div`
 
 const CardCVCLabel = styled.label`
   font-size: 12px;
-`;
-
-const CardCVCInput = styled.input<ErrorFlag>`
-  border: solid 1px ${(props) => (props.isError ? "#FF3D3D" : "#acacac")};
-  border-radius: 2px;
-  padding: 0.5rem;
-  font-size: 11px;
-  height: 32px;
-  -moz-appearance: textfield;
-  &::placeholder {
-    color: #acacac;
-  }
-
-  &::-webkit-inner-spin-button,
-  &::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
 `;
