+++ app/src/components/form/CardExpiryDateInput.tsx
@@ -1,4 +1,4 @@
-import styled from "@emotion/styled";
+import { CardInput, CardFieldset, CardLegend } from "../../style/CardInput";
 import { Validator } from "../../validators/CardValidator";
 import { useState } from "react";
 import { ErrorMessage } from "./ErrorMessage";
@@ -53,9 +53,9 @@ export function CardExpiryDateInputContainer({
 
   return (
     <>
-      <CardExpiryDateFieldset>
-        <CardExpiryDateLegend>유효기간</CardExpiryDateLegend>
-        <CardExpiryDateInput
+      <CardFieldset>
+        <CardLegend>유효기간</CardLegend>
+        <CardInput
           type="text"
           inputMode="numeric"
           id="expiry-month"
@@ -66,7 +66,7 @@ export function CardExpiryDateInputContainer({
           isError={isError["expiry-month"].state}
           placeholder="MM"
         />
-        <CardExpiryDateInput
+        <CardInput
           type="text"
           inputMode="numeric"
           id="expiry-year"
@@ -77,44 +77,8 @@ export function CardExpiryDateInputContainer({
           isError={isError["expiry-year"].state}
           placeholder="YY"
         />
-      </CardExpiryDateFieldset>
+      </CardFieldset>
       <ErrorMessage message={isError["message"]} />
     </>
   );
 }
-
-type ErrorFlag = {
-  isError?: boolean;
-};
-
-const CardExpiryDateFieldset = styled.fieldset`
-  border: none;
-  display: flex;
-  gap: 0.625rem;
-  padding: 0;
-  margin: 0;
-`;
-
-const CardExpiryDateLegend = styled.legend`
-  font-size: 12px;
-  margin: 8px 0;
-`;
-
-const CardExpiryDateInput = styled.input<ErrorFlag>`
-  border: solid 1px ${(props) => (props.isError ? "#FF3D3D" : "#acacac")};
-  border-radius: 2px;
-  padding: 0.5rem;
-  font-size: 11px;
-  width: 100%;
-  height: 32px;
-  -moz-appearance: textfield;
-  &::placeholder {
-    color: #acacac;
-  }
-
-  &::-webkit-inner-spin-button,
-  &::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-`;
